Re-structure test app validators and yml file (#227)
