+++ test/app_testing/run_example_tests.sh
@@ -1,4 +1,5 @@
+#!/usr/bin/env bash
 
 pip install tensorflow
 
-python test_runner.py --poc ../../nvflare/poc --n_clients 2 --yaml test_apps/test_examples.yml --app_path ../../examples --cleanup
+python test_runner.py --poc ../../nvflare/poc --n_clients 2 --yaml test_examples.yml --app_path ../../examples --cleanup
