--- test/app_testing/test_apps/test_examples.yml
@@ -1,20 +1,0 @@
-
-
-
-tests:
-  - app_name: hello-pt
-    validators:
-      - test.app_testing.test_apps.validators.pt_model_validator.PTModelValidator
-  - app_name: hello-numpy-cross-val
-    validators:
-      - test.app_testing.test_apps.validators.cross_val_result_validator.CrossResultValidator
-      - test.app_testing.test_apps.validators.sag_result_validator.SAGResultValidator
-  - app_name: hello-numpy-sag
-    validators:
-      - test.app_testing.test_apps.validators.sag_result_validator.SAGResultValidator
-  - app_name: hello-tf2
-    validators:
-      - test.app_testing.test_apps.validators.tf_model_validator.TFModelValidator
-  - app_name: hello-cyclic
-    validators:
-      - test.app_testing.test_apps.validators.tf_model_validator.TFModelValidator
